mv generate_samples_batch.py to synthesize.py
